+++ shared/src/androidMain/kotlin/dcapi/WalletAPIData.kt
@@ -1,4 +1,4 @@
-package digitalcredentialsapi
+package dcapi
 
 import android.content.Intent
